refactor(brIBMr6C): move unit test config
Also add new method: getTestAuthoritiesDTOs()
